Fail gracefully for remote peptide search jobs.
